improve interface
