+++ src/cryojax/rotations/__init__.py
@@ -1,5 +1,2 @@
-from ._rotation import (
-    AbstractRotation as AbstractRotation,
-    AbstractMatrixLieGroup as AbstractMatrixLieGroup,
-    SO3 as SO3,
-)
+from ._rotation import AbstractRotation as AbstractRotation
+from ._lie_group import AbstractMatrixLieGroup as AbstractMatrixLieGroup, SO3 as SO3
